refactor: Seat 엔티티와 연관 타입인 SeatType을 별도의 도메인 패키지로 이동
