+++ src/main/java/com/jvnlee/catchdining/domain/seat/model/SeatType.java
@@ -1,0 +1,5 @@
+package com.jvnlee.catchdining.domain.seat.model;
+
+public enum SeatType {
+    HALL, ROOM, BAR
+}
