--- src/main/java/com/jvnlee/catchdining/entity/SeatType.java
@@ -1,5 +1,0 @@
-package com.jvnlee.catchdining.entity;
-
-public enum SeatType {
-    HALL, ROOM, BAR
-}
